### 2019-01-25 #280 @eebrown
Submission: tacmagic - PET Analysis in R
Submitting Author: Eric Brown (@eebrown)
Repository:   https://github.com/eebrown/PET
Version submitted:   0.1.9
Editor: Melina Vidoni (@melvidoni)
Reviewer 1: Jon Clayden (@jonclayden)
Reviewer 2: Brandon Hurr (@bhive01)
Version accepted: TBD

---

```
Package: tacmagic
Title: tacmagic: PET Analysis in R
Version: 0.1.9
Authors@R: c(person("Eric", "Brown",
                    email = "eb@ericebrown.com",
             person("Ariel", "Graff-Guerrero",
                    role = c("dgs")))
Description: To faciliate analysis of positron emission tomography (PET) time
    activity curve (TAC) data, and to encourage open science and replicability,
    this package supports data loading and analysis of multiple TAC file
    formats. Functions are available to analyze loaded tac data for individual
    participants or in batches. Major functionality includes weighted TAC
    merging by region of interest (ROI), calculating models including SUVR and
    DVR, basic plotting functions and calculation of cut-off values. Please see
    the walkthrough vignette for a detailed overview of tacmagic functions.
    graphics,
    grDevices,
    pracma,
    R.matlab
    covr
```

Please see the presubmission inquiry as this package was deemed in scope.(#277)

This packages supports loading positron emission tomography (PET) time-activity curves (TACs) from several different major formats, to allow a common subsequent analysis pipeline within R ("data extraction"). There are functions to support merging of regional time-activity curves, an often important analysis step ("data munging"). Additional analysis is available in the package, which virtue of being open source promotes open science reproducibility.

Scientists working with PET data needing to analyze TAC data in R can take advantage of the loading (including batch loading) functions and basic plotting. Those who need basic non-invasive models can take advantage of the implemented models. We will invite contributors to extend the package in its support for file formats and models.

There do not appear to be packages that offer the loading and merging functions. There are packages that have some PET models implemented but not full overlap.

Issue #277 - @sckott
