+++ R/saving.R
@@ -29,7 +29,7 @@ save_tac <- function(tac, outfile) {
     names(tac)[1] <- paste0("start[", attributes(tac)$time_unit, "]")
 	names(tac)[2] <- paste0("end[", attributes(tac)$activity_unit, "]")
 
-	write.table(tac, file = outfile, append = FALSE, quote = TRUE, sep = "\t",
+	write.table(tac, file = outfile, append = FALSE, quote = FALSE, sep = "\t",
                  eol = "\n", na = "NaN", dec = ".", row.names = FALSE,
                  col.names = TRUE, qmethod = c("escape", "double"),
                  fileEncoding = "")
