### 2019-01-28 #280 @melvidoni
### Editor checks:

---

#### Editor comments

Thank you for your submission @eebrown! The output from [goodpractice](https://github.com/MangoTheCat/goodpractice/) is perfect. I will start looking for reviewers.

```
Preparing: covr
Preparing: cyclocomp
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck

♥ Mhm! Wonderful package! Keep up the majestic work!
```

---

Reviewers: TBD
Due date:
