### 2019-02-12 d49a41a
typo in loading.R
