+++ R/loading.R
@@ -24,7 +24,7 @@
 #'   \item "voistat": PMOD TAC .voistat files used in combination with PMOD 
 #'          .acqtimes file for start/stop times.
 #'   \item "magia": magia pipeline .mat tac file
-#'   \item "DFF": Turku PET Centre's DFT format
+#'   \item "DFT": Turku PET Centre's DFT format
 #' }
 #' 
 #'@export
