### 2019-02-12 cdaaa43
Improved tac loading: numeric checking, more unit formats, and flexibility with DFT files
