### 2019-02-12 #280 @eebrown
Hi again @bhive01. I've found some .tac and .dft sample files [sample_tac_files.zip](https://github.com/ropensci/software-review/files/2856307/sample_tac_files.zip) from the open source PET analysis C library from the Turku PET centre (www.turkupetcentre.net). I have had to make some minor changes to the tac loading functions to support some idiosyncrasies with the files, so I have updated the devel branch with that. I am not sure what the best practice is for updating the master branch while under review, but I can merge in these changes if that's best (https://github.com/eebrown/PET/compare/devel)

Please let me know if there is something else you were hoping to test more.
