### 2019-02-12 #280 @melvidoni
@eebrown If the changes are not too fundamental (only for the examples), you can then merge the changes once you have received the reviews.
