### 2019-02-12 #280 @bhive01
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Firstly, while I am very much an image analyst, "I don't do humans" (Ace Ventura 1994), so I'm a bit out of my element with PET image model outputs. For that purpose, and as far as I can tell, this package works as intended and provides meaningful data processing and plots for examining PET data output in the various formats built-in (PMOD TAC, Voistat, DFT, Magia/Matlab).
Given the sample data, exported functions with examples all work as described.

Below are the outputs from goodpractice, devtools::check, and devtools::test. Overall there are a few issues found by goodpractice::gp around unit testing, but coverage is pretty good.

```
devtools::check("~/PET")
Updating tacmagic documentation
Writing NAMESPACE
Loading tacmagic
Writing NAMESPACE
── Building ────────────────────────────────────────────────────────────────────────────────────────────────────────────── tacmagic ──
Setting env vars:
● CFLAGS    : -Wall -pedantic
● CXXFLAGS  : -Wall -pedantic
● CXX11FLAGS: -Wall -pedantic
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
─  building ‘tacmagic_0.1.9.tar.gz’ation .../DESCRIPTION’ ...
✔  checking for file ‘tacmagic/DESCRIPTION’
─  checking extension type ... Package
─  this is package ‘tacmagic’ version ‘0.1.9’
─  package encoding: UTF-8
✔  checking package namespace information ... checking for file ‘tacmagic/DESCRIPTION’
✔  checking package dependencies (905ms)
✔  checking for hidden files and directoriesms)
✔  checking for portable file names
✔  checking for sufficient/correct file permissions ...
✔  checking serialization versions
✔  checking top-level filesa-information ...e installed ...
✔  checking for left-over files
✔  checking foreign function calls (339ms)...aded cleanly ...ependencies ...
✔  checking contents of ‘data’ directory examples ...
✔  checking installed files from ‘inst/doc’ saves ...
── R CMD check results ─────────────────────────────────────────────────────────────────────────────────────────── tacmagic 0.1.9 ────
Duration: 21.9s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
──────────────────────────────────────────────────────────── tacmagic 0.1.9 ────
```
```
devtools::test("~/PET")
Loading tacmagic
Testing tacmagic
✔ | 24       | TAC data loading and weighted averages [0.6 s]

══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════

OK:       104
 [0.6 s]
```

```
library(goodpractice)
gp("~/PET")
Preparing: covr
Preparing: cyclocomp
   checking vignette meta-information ......4/q75sn_4526x2l0cj6jc6c4mm0000gn/T/RtmpZ2PERy/remotes275767f4aa2/tacmagic/DESCRIPTION’ ...
* installing *source* package ‘tacmagic’ ...
** R
** data
*** moving datasets to lazyload DB
** inst
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
** building package indices
** installing vignettes
** testing if installed package can be loaded
* DONE (tacmagic)
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck
── GP tacmagic ──────────────────────────────────────────────────────────────────────────────────────────────────────────

    R/batches.R:65:NA
    R/loading_utilities.R:80:NA
    R/loading_utilities.R:150:NA
    R/loading_utilities.R:151:NA
    R/loading_utilities.R:178:NA

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no redirection of stdout/stderr. Hmm ... looks like a package You may want to clean up by 'rm -Rf
    /var/folders/p4/q75sn_4526x2l0cj6jc6c4mm0000gn/T//Rtmp3exf2k/Rd2pdf5f260a515b8'
─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

### Function Naming

You have functions named after authors of papers (Logan and Aiz). I guess this is a good idea because it tells you where they came from, but perhaps they could have more generic names so in the future you could use a "method" argument if there are other methods added in the future. This suggestion comes from a place of naiveté so _cum grano salis_.

```
ls(getNamespace("tacmagic"), all.names = TRUE)
 [1] ".__NAMESPACE__."       ".__S3MethodsTable__."  ".packageName"          "batch_load"
 [5] "batch_tm"              "batch_voistat"         "checkACQtimes"         "compare_tac_form"
 [9] "cutoff_aiz"            "DVR_all_ref_Logan"     "DVR_ref_Logan"         "find_t_star"
[13] "load_acq"              "load_header_DFT"       "load_ROIs_DFT"         "load_tac"
[17] "load_tac_DFT"          "load_tac_magia"        "load_tac_PMOD"         "load_tac_voistat"
[21] "load_tacs"             "load_voistat"          "load_vol"              "load_vol_DFT"
[25] "model_batch"           "model_definitions"     "new_table"             "plot_ref_Logan"
[29] "plot_tac"              "pos_anyroi"            "ref_Logan_lm"          "ref_Logan_xy"
[33] "references"            "roi_ham_full"          "roi_ham_pib"           "roi_ham_stand"
[37] "save_tac"              "suvr"                  "suvr_auc"              "tac_roi"
[41] "upper_inner_fence"     "validate_tac"          "validateTACvoistat"    "vAUC"
[45] "vintegrate"            "volumesFromBPndPaste"  "volumesFromVoistatTAC"
```
Some inconsistencies in the naming convention used for functions with the use of CamelCase. https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming suggests using _snake_case_ which is mostly used. I would update the following:

* checkACQtimes
* validateTACvoistat
* vintegrate
* volumesFromBPndPaste
* validateTACvoistat
* volumesFromVoistatTAC
* vAUC

Where the naming is applied it’s clear from the name what the function is responsible for.

### `loading.R` and `loading_utilities.R`
Load functions serve to validate inputs and contort output files into a TAC-like format. TAC format is a very wide data.frame. Does not match "tidy" principles, but that's ok.

Might want to include appropriate units for importing from a matlab file in the example. This is done in the vignette.
```
f_magia <- system.file("extdata", "AD06_tac_magia.mat", package="tacmagic")
load_tac(f_magia, format = "magia", time_unit = "seconds", activity_unit = "kBq/cc")`
```

DFT is misspelled on line 27 as DFF in the example in `loading.R`

How fragile are these naming conventions?
> Asked author for more test cases and some of these was tested. Changes already being made.

Would it make sense to just read in the volume information as the main function? (at least for voistat because it's included already in the file) Also applicable for DFT AFICT. not for the BPND (Binding Potential BPnd) though, because it's a separate exported CSV file.

I wonder if instead of inputting the format, you could use the input extension as "default". Would maybe make that more simple for the end user. Default being PMOD tac is probably a good one as that seems like the most common (hard to tell from Google really).

`compare_tac_form()` looks like it's in the wrong place in loading_utilities.R
it's not used in any of the loading functions, but is used only by the plot_tac() in tac.R
Seems like it should be moved there to me.

### `tac.R`
functions are all about combining ROIs using the tissue definitions.
	Uses weighted means to summarize based upon the tissue volume as contribution

I notice there are a few instances where you are applying attributes to a munged file. Perhaps `apply_attributes()` could be a function that does this. This could sit in the `utilities.R` file.

#### `tac.R` Plotting
I feel like perhaps the plotting functions should be separated from the TAC ROI combination function. Put `compare_tac_form()` and `plot_tac` into a `plot.R` file (or whatever you want to call it). `plot_ref_logan()` from `logan.R`

line 121-122 you're using Rainbow (and restricting to red -- green). Concern is that rainbow produces color ramps not good for color blindness.

I plotted everything once and it got very busy.
`plot_tac(AD06, ROIs=names(AD06), time="minutes", title="PIB time activity curves for AD06")`
I wonder if a warning is worth including if above a certain number of ROIs you can't really tell the difference between similar colors. Not critical.

### `suvr.R`
Two functions that compute standardized uptake value ratios (and their AUC)
They are well-described and have examples. The examples work. I do not fully understand the meaning of their outputs, but things work as expected. I notice there are a few #TODOs around testing inputs are numeric. These functions could definitely use some love in this area.

### `logan.R`
Functions work as described. Plot function could be removed and put into a plotting file?

### `cutoff.R`
DVR is not defined (Distribution volume ratios) in the function descriptions (is in vignette)
PiB is not defined (Pittsburg Compound B?) in function descriptions or vignette
usage is in the walkthrough, but no examples for the help pages

### `batches.R` and `batch_models.R`
Describe methods of importing and analyzing many TAC models in series (automated). Would be useful to have the example in the help file in the vignette. The generic example in the vignette is not runable.

### `ROI_definitions.R`
This file looks good, is well described and everything returns a list of ROIs that is used by `tac_roi.r`

### `saving.R`
The NA values are being stored as NaN on save. I assume this is the tac default?

It seems if you load volumes that these are stuck in the voistat or DFT files.

### `tacmagic.R`
This file  allows you to do ?tacmagic and see the exported functions of the package (i love this)

### `references.R`
References look in order

### `data.R`
shows how the fake data for testing the batch processing was made. clear.
