### 2019-02-13 f1ac29e
updated remaining camelCase function names to snake_case for consistency
