### 2019-02-14 b03af39
Added extension checking to file loading functions--checks file extension matches specified format and warns if not.
