+++ NAMESPACE
@@ -34,6 +34,7 @@ importFrom(stats,lm)
 importFrom(stats,quantile)
 importFrom(stats,sd)
 importFrom(stats,weighted.mean)
+importFrom(tools,file_ext)
 importFrom(utils,head)
 importFrom(utils,read.csv)
 importFrom(utils,read.delim)
