### 2019-02-14 c9ded13
moved compare_tac_form() from loading_utilities.R to tac.R where it is used.
