+++ R/loading_utilities.R
@@ -285,17 +285,3 @@ load_ROIs_DFT <- function(header) {
 
   return(paste(vars, secvars, sep="_"))
 }
-
-
-# Used by the plot, or any function that needs 2 tacs, to ensure their overall
-# form and attributes are equal (except the ROIs)
-compare_tac_form <- function(tac, tac2) {
-  if (!all.equal(tac$start, tac2$start)) stop("tac start times not equal")
-  if (!all.equal(tac$end, tac2$end)) stop("tac end times not equal")
-  a1 <- attributes(tac)
-  a2 <- attributes(tac2)
-  if (!all.equal(a1$time_unit, a2$time_unit)) stop("tac time units not equal")
-  if (!all.equal(a1$activity_unit, a2$activity_unit)) stop("tac start times 
-                                                            not equal")
-  return(TRUE) 
-}
