### 2019-02-14 434b37c
in tac_roi() moved attributes assignment to a helper function copy_tac_attributes()
