+++ R/utilities.R
@@ -13,3 +13,13 @@ new_table <- function(tac, varname="VALUE") {
   names(VALUEtable) <- varname
   return(VALUEtable)
 }
+
+#' Copies the time and activity unit attributes from origin to destination;
+#' and assigns tm_type as "tac"
+#'@noRd
+copy_tac_attributes <- function(origin, destination) {
+  attributes(destination)$time_unit <- attributes(origin)$time_unit
+  attributes(destination)$activity_unit <- attributes(origin)$activity_unit
+  attributes(destination)$tm_type <- "tac"
+  return(attributes(destination))
+}
