### 2019-02-14 7d2b4b1
moved plotting functions to plot.R
