+++ R/plot.R
@@ -1,0 +1,154 @@
+##################################
+## tacmagic - PET Analysis in R ##
+## tac.R                        ##
+## (C) Eric E. Brown  2018      ##
+## Beta version--check all work ##
+##################################
+
+
+#' Plots time activity curves from 1 or 2 participants or groups.
+#'
+#'@export
+#'@param TACtable1 (e.g. from tac_roi() or load_tac())
+#'@param TACtable2 An optional, second TAC, to plot for comparison
+#'@param ROIs A vector of ROIs to plot, names matching the TAC headers
+#'@param ymax The maximum value on the y-axis
+#'@param time "seconds" or "minutes" depending on desired x-axis, converts tac
+#'@param title A title for the plot
+#'@return Creates a plot
+#'@family tac functions 
+#'@examples
+#' # f_raw_tac and f_raw_vol are the filenames of PMOD-generated files
+#' f_raw_tac <- system.file("extdata", "AD06.tac", package="tacmagic") 
+#' f_raw_vol <- system.file("extdata", "AD06_TAC.voistat", package="tacmagic")
+#' 
+#' tac <- load_tac(f_raw_tac)
+#' vol <- load_vol(f_raw_vol)
+#' AD06_tac_nc <- tac_roi(tac, vol, roi_ham_full(), merge=FALSE, PVC=FALSE)
+#' plot_tac(AD06_tac_nc, ROIs=c("frontal", "cerebellum"), title="Example Plot")
+plot_tac <- function(TACtable1, TACtable2=NULL, ROIs, ymax=25, 
+                     time="minutes", title="") {
+  
+  if (!all(ROIs %in% names(TACtable1))) stop("ROIs are not in TACtable1")
+
+  if (!validate_tac(TACtable1)) stop("The 1st tac object did not validate.")
+  if (!is.null(TACtable2)) {
+    if (!validate_tac(TACtable2)) stop("The 2nd tac object did not validate.")
+  }
+  
+  if (!(time %in% c("seconds", "minutes"))) stop("Time must be \"seconds\" or 
+                                                 \"minutes\"")
+
+  if (time == "minutes") {
+    if (attributes(TACtable1)$time_unit == "minutes") time_conversion <- 1
+    if (attributes(TACtable1)$time_unit == "seconds") time_conversion <- 60
+  }
+
+  if (time == "seconds") {
+    if (attributes(TACtable1)$time_unit == "seconds") time_conversion <- 1
+    if (attributes(TACtable1)$time_unit == "minutes") time_conversion <- 1/60
+  }
+
+  # Sets up the plot using the frame start from the TAC file for the x axis
+  # and converting to minutes if chosen. 
+   
+  plot(1,type='n',
+       xlim=c(TACtable1$start[1],
+              TACtable1$start[length(TACtable1$start)]/time_conversion),
+              ylim=c(0,ymax),
+              xlab=paste0("Time (", time, ")"),
+              ylab=paste0("Activity (", 
+                          attributes(TACtable1)$activity_unit, ")"),
+              main=title)
+  
+  # Separate colour ranges for each group of TACs
+  colour1 <- rainbow(length(ROIs), start=0, end=0.25)
+  colour2 <- rainbow(length(ROIs), start=0.5, end=0.8)
+  
+  # Plots the ROIs as specified in the ROIs argument
+  for (ROI in seq_along(ROIs)) {
+
+    lines(x=TACtable1$start/time_conversion, 
+          y=TACtable1[,ROIs[ROI]], 
+          type='o', 
+          col=colour1[ROI], 
+          lwd=2)
+    
+    # Only if 2nd TAC table is provided, plots second on same plot
[…]
