+++ man/plot_ref_Logan.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/logan.R
+% Please edit documentation in R/plot.R
 \name{plot_ref_Logan}
 \alias{plot_ref_Logan}
 \title{Non-invasive reference Logan plot}

[…]
