### 2019-02-14 6806094
function validate_suvr_params() to check that supplied params for suvr() or suvr_auc() are valid.
