### 2019-02-14 efe3250
update cutoff_aiz() documentation and example
