### 2019-02-15 629712a
Added function split_pvc() and clarified batch examples in vignette and documentation. Clarified batch arguments.
