+++ R/tac.R
@@ -59,3 +59,34 @@ tac_roi <- function(tac, volumes, ROI_def, merge, PVC) {
 
   return(calculated_TACs)
 }
+
+#' Subset PMOD tacs with or without PVC
+#' 
+#' When partial volume correction (PVC) is used in PMOD, the saved tac files
+#' have ROIs with and without PVC. When loaded with load_tac()) it may be 
+#' desirable to keep only either the PVC or non-PVC tacs. This returns a tac 
+#' object that is a subset of the input tac object with only the PVC or non-PVC
+#' tacs. This relies on PMOD's convention of labelling tac columns with "_C".
+#'
+#'@export
+#'@param tac The time-activity curve data from loading function (PMOD)
+#'@param PVC If TRUE, includes columns with "_C", if FALSE, ones without "_C"
+#'@family tac functions 
+#'@return Time-activity curve object
+#'@examples 
+#' # f_raw_tac and f_raw_vol are the filenames of PMOD-generated files
+#' f_raw_tac <- system.file("extdata", "AD06.tac", package="tacmagic") 
+#' 
+#' tac <- load_tac(f_raw_tac)
+#' tac_pvc <- split_pvc(tac, TRUE)
+#' tac_nc <- split_pvc(tac, FALSE)
+split_pvc <- function(tac, PVC=TRUE) {
+
+  validate_tac(tac)
+  subset <- cbind(tac[1:2], tac[endsWith(names(tac), "_C") == PVC])
+  attributes(subset) <- copy_tac_attributes(tac, subset)
+  validate_tac(subset)
+
+  return(subset)
+
+}
