+++ doc/walkthrough.R
@@ -94,6 +94,18 @@ AD06_DVR
 
 
 ## ------------------------------------------------------------------------
+participants <- c(system.file("extdata", "AD06.tac", package="tacmagic"),
+                   system.file("extdata", "AD07.tac", package="tacmagic"),
+                   system.file("extdata", "AD08.tac", package="tacmagic"))
+ 
+tacs <- batch_load(participants, tac_file_suffix="", PVC=FALSE)
+ 
+batch <- batch_tm(tacs, models=c("SUVR", "Logan"), ref="Cerebellum_r",
+                  SUVR_def=c(3000,3300,3600), k2prime=0.2, t_star=23)
+
+
+
+## ------------------------------------------------------------------------
 fake_DVR[1:5,]
 
 
