+++ man/tac_roi.Rd
@@ -34,6 +34,6 @@ AD06_tac_nc <- tac_roi(tac, vol, roi_ham_full(), merge=FALSE, PVC=FALSE)
 }
 \seealso{
 Other tac functions: \code{\link{plot_tac}},
-  \code{\link{save_tac}}
+  \code{\link{save_tac}}, \code{\link{split_pvc}}
 }
 \concept{tac functions}
