+++ NAMESPACE
@@ -17,6 +17,7 @@ export(roi_ham_full)
 export(roi_ham_pib)
 export(roi_ham_stand)
 export(save_tac)
+export(split_pvc)
 export(suvr)
 export(suvr_auc)
 export(tac_roi)

[…]
