### 2019-02-15 #280 @eebrown
# Response to @bhive01

Thank you again for taking the time to thoroughly review the package. I have implemented many of your suggestions on the devel branch as indicated below. I noticed you skipped the JOSS section -- I am not sure if this is done later in the review, but I would like to submit to JOSS after the reviews.

> Firstly, while I am very much an image analyst, "I don't do humans" (Ace Ventura 1994), so I'm a bit out of my element with PET image model outputs. For that purpose, and as far as I can tell, this package works as intended and provides meaningful data processing and plots for examining PET data output in the various formats built-in (PMOD TAC, Voistat, DFT, Magia/Matlab).
> Given the sample data, exported functions with examples all work as described.
>
> Below are the outputs from goodpractice, devtools::check, and devtools::test. Overall there are a few issues found by goodpractice::gp around unit testing, but coverage is pretty good.

 [...]

>   ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd problems.
>   ✖ fix this R CMD check ERROR: Re-running with no redirection of stdout/stderr. Hmm ... looks like a package You may want to clean up by 'rm -Rf
>     /var/folders/p4/q75sn_4526x2l0cj6jc6c4mm0000gn/T//Rtmp3exf2k/Rd2pdf5f260a515b8'
> ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
> ```

I am not sure what to make of the WARNING and ERROR you see with goodpractice, as I don't get those errors on my Ubuntu or macOS machine.

> ### Function Naming
> You have functions named after authors of papers (Logan and Aiz). I guess this is a good idea because it tells you where they came from, but perhaps they could have more generic names so in the future you could use a "method" argument if there are other methods added in the future. This suggestion comes from a place of naiveté so _cum grano salis_.
>
> ```
> ls(getNamespace("tacmagic"), all.names = TRUE)
>  [1] ".__NAMESPACE__."       ".__S3MethodsTable__."  ".packageName"          "batch_load"
>  [5] "batch_tm"              "batch_voistat"         "checkACQtimes"         "compare_tac_form"
>  [9] "cutoff_aiz"            "DVR_all_ref_Logan"     "DVR_ref_Logan"         "find_t_star"
> [13] "load_acq"              "load_header_DFT"       "load_ROIs_DFT"         "load_tac"
> [17] "load_tac_DFT"          "load_tac_magia"        "load_tac_PMOD"         "load_tac_voistat"
> [21] "load_tacs"             "load_voistat"          "load_vol"              "load_vol_DFT"
> [25] "model_batch"           "model_definitions"     "new_table"             "plot_ref_Logan"
> [29] "plot_tac"              "pos_anyroi"            "ref_Logan_lm"          "ref_Logan_xy"
> [33] "references"            "roi_ham_full"          "roi_ham_pib"           "roi_ham_stand"
> [37] "save_tac"              "suvr"                  "suvr_auc"              "tac_roi"
> [41] "upper_inner_fence"     "validate_tac"          "validateTACvoistat"    "vAUC"
> [45] "vintegrate"            "volumesFromBPndPaste"  "volumesFromVoistatTAC"
> ```
> Some inconsistencies in the naming convention used for functions with the use of CamelCase. https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming suggests using _snake_case_ which is mostly used. I would update the following:
>
> * checkACQtimes
> * validateTACvoistat
> * vintegrate
> * volumesFromBPndPaste
> * validateTACvoistat
> * volumesFromVoistatTAC
> * vAUC
>
> Where the naming is applied it’s clear from the name what the function is responsible for.

Thanks for pointing this out. I've now renamed the above functions to be consistent with the naming guidelines and the rest of the package's functions. As for the author-named functions, I think I will leave as-is currently, as I could always add a more generic function after (but harder to go other way around, and it's not yet clear what parameters the generic function would need).

Commit f1ac29e.

> ### `loading.R` and `loading_utilities.R`
> Load functions serve to validate inputs and contort output files into a TAC-like format. TAC format is a very wide data.frame. Does not match "tidy" principles, but that's ok.
>
> Might want to include appropriate units for importing from a matlab file in the example. This is done in the vignette.
>
> ```
> f_magia <- system.file("extdata", "AD06_tac_magia.mat", package="tacmagic")
> load_tac(f_magia, format = "magia", time_unit = "seconds", activity_unit = "kBq/cc")`
> ```
> DFT is misspelled on line 27 as DFF in the example in `loading.R`
>

Thanks. Fixed in d49a41a.

> How fragile are these naming conventions?
>
> > Asked author for more test cases and some of these was tested. Changes already being made.
>
> Would it make sense to just read in the volume information as the main function? (at least for voistat because it's included already in the file) Also applicable for DFT AFICT. not for the BPND (Binding Potential BPnd) though, because it's a separate exported CSV file.
>
> I wonder if instead of inputting the format, you could use the input extension as "default". Would maybe make that more simple for the end user. Default being PMOD tac is probably a good one as that seems like the most common (hard to tell from Google really).
>

I like the idea of making use of the file extensions in data loading. I think it is a balance of convenience vs. code safety and I think for now it makes sense to err on code safety. So instead of assigning the format from the file extension, I have added the file extension as a check that now warns the user if there is a mismatch between their specified file type versus the file extension. Commit b03af39.

> `compare_tac_form()` looks like it's in the wrong place in loading_utilities.R
> it's not used in any of the loading functions, but is used only by the plot_tac() in tac.R
> Seems like it should be moved there to me.

Agreed. Commit c9ded13.

> ### `tac.R`
> functions are all about combining ROIs using the tissue definitions.
> Uses weighted means to summarize based upon the tissue volume as contribution
>
> I notice there are a few instances where you are applying attributes to a munged file. Perhaps `apply_attributes()` could be a function that does this. This could sit in the `utilities.R` file.

I made a function called copy_tac_attributes() to get the specific attributes from an original tac object and assign them to the copy. This will make things more convenient if more attributes are needed in the future. Commit 434b37c.

> #### `tac.R` Plotting
> I feel like perhaps the plotting functions should be separated from the TAC ROI combination function. Put `compare_tac_form()` and `plot_tac` into a `plot.R` file (or whatever you want to call it). `plot_ref_logan()` from `logan.R`
>
> line 121-122 you're using Rainbow (and restricting to red -- green). Concern is that rainbow produces color ramps not good for color blindness.
>
> I plotted everything once and it got very busy.
> `plot_tac(AD06, ROIs=names(AD06), time="minutes", title="PIB time activity curves for AD06")`
> I wonder if a warning is worth including if above a certain number of ROIs you can't really tell the difference between similar colors. Not critical.

I have moved the plotting functions into plot.R as suggested. As for the warnings about plotting too many ROIs, I think it may not be necessary as it would be self evident. I see the plotting function currently as a quick way to visualize tacs when performing analyses rather than for the generation of publication-ready plots, but this is obviously an area that could be developed. I think at that point it will make sense to overhaul the visuals including a better colour scheme, among other optimizations. Commit 7d2b4b1.

> ### `suvr.R`
> Two functions that compute standardized uptake value ratios (and their AUC)
> They are well-described and have examples. The examples work. I do not fully understand the meaning of their outputs, but things work as expected. I notice there are a few #TODOs around testing inputs are numeric. These functions could definitely use some love in this area.

I created a new function validate_suvr_params() in suvr.R that handles the checks that were needed and is now used in both suvr() and suvr_auc(). Commit 6806094.

> ### `logan.R`
> Functions work as described. Plot function could be removed and put into a plotting file?

Done as previously noted.

> ### `cutoff.R`
> DVR is not defined (Distribution volume ratios) in the function descriptions (is in vignette)
> PiB is not defined (Pittsburg Compound B?) in function descriptions or vignette
> usage is in the walkthrough, but no examples for the help pages

Fixed. Commit efe3250.

> ### `batches.R` and `batch_models.R`
> Describe methods of importing and analyzing many TAC models in series (automated). Would be useful to have the example in the help file in the vignette. The generic example in the vignette is not runable.

I have now added a runable example to the vignette. I was avoiding it, because the batch functions rely on real-world file names and directory structures but I have hopefully made it clear how it would be different in practice. In the process I also added a new function that is useful when PMOD tacs are loaded but when combining ROIs is not desired: split_pvc(). It is useful for PMOD tac files that have both PVC-corrected and uncorrected ROIs but when only one copy is needed. It is in the tac.R file and documented. Commit 629712a. Rebuilt vignette commit b9ce86d.

> ### `ROI_definitions.R`
> This file looks good, is well described and everything returns a list of ROIs that is used by `tac_roi.r`
>
> ### `saving.R`
> The NA values are being stored as NaN on save. I assume this is the tac default?

Yes this is for consistency with PMOD files.

> It seems if you load volumes that these are stuck in the voistat or DFT files.

Yes the PMOD tac files don't have volume information and saving is currently built around PMOD .tac files.

> ### `tacmagic.R`
> This file allows you to do ?tacmagic and see the exported functions of the package (i love this)

Thanks :)

> ### `references.R`
> References look in order
>
> ### `data.R`
> shows how the fake data for testing the batch processing was made. clear.

Please let me know if you have any questions/concerns about the changes.
Best,
Eric
