### 2019-02-18 #280 @jonclayden
>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Thanks for the opportunity to review this well-constructed and well-documented package for analysing time–activity curve (TAC) data derived from PET. The package appears to follow the requirements of the rOpenSci packaging guidelines, `R CMD check` and the package tests run fine on my system, and test coverage is very good at 93%. I have a few comments about the vignette and the code interface, which are laid out below. I hope they're helpful.

(NB: I have not read the other review, in the interests of objectivity, so feel free to disregard any comments that have already been addressed!)

1. I work with MRI, not PET, so please forgive any misunderstanding on my part, but from reading the vignette it seems like TAC analysis is only part of a full PET analysis pipeline. A certain amount of external software is mentioned, which presumably carries out preprocessing, segmentation, coregistration with MRI and so on. I think it would provide useful context and clarify the package's role if you could add a diagram or textual outline of the full pipeline early in the vignette, with a clear indication of which parts are handled by `tacmagic`. I felt like I had a picture of this in my head by the end of the document, but it would have been helpful to have it made explicit early on.
2. The core data structure in the package is a data frame with certain attributes and obligatory columns. I wonder whether it would be worth giving this an (informal, "S3") class, say `"tac"`, for clarity and to give you to chance to exploit generic functions like `plot()` for more concise usage. You could also consider a custom `print()` method which summarises the data, to help the user with some sanity-checking, e.g. number of time periods, period length, number of ROIs, etc. — but that's entirely up to you.
3. I would suggest giving the user a bit more control over standard plot characteristics, particularly colours. The rainbow palette may be hard to interpret for a user with colour vision deficiency (colour-blindness). You could approach this by adding an ellipsis (`...`) argument to `plot_tac()` and `plot_ref_Logan()`, and passing it on to one of the base graphics primitive functions you use.
4. Consider also using an ellipsis argument to `batch_tm()`, rather than piling up arguments from `suvr()` and `DVR_all_ref_Logan()`.
5. The function names `DVR_ref_Logan` and `DVR_all_ref_Logan` seem very verbose compared to others in the package, and the need for two separate functions isn't clear to me. Couldn't a putative `dvr()` function take a target argument but return all (like `suvr()`) if no specific targets are specified?
6. Having model-fitting parameters among the arguments to `plot_ref_Logan()` feels like an unnecessary convolution of its role. Perhaps you could arrange for the SUVR and DVR functions to return a data type that can be plotted without repeating the fit?
7. I would suggest adding reference information to the function documentation using the `@references` tag, and removing the `references()` function.
8. In the vignette the syntax `help(load_tac())` is used, but this doesn't work. The function shouldn't be called when passing it to `help()`.
9. There are a number of typos dotted through the package: exctracted, inntegration, cerbellum, caluclation, arguements, conver, BBy.
