### 2019-02-19 cd1fecf
moved references from dedicated file/function references() to function documentation as appropriate
