+++ R/cutoff.R
@@ -8,16 +8,16 @@
 
 #' Cutoff value caluclation using method described in Aizenstein et al. 2008
 #' 
-#' See references()$Aizenstein and the tacmagic walkthrough vignette. The 
-#' authors proposed a standardized method of calculating Pittsburg Compound B 
-#' (PIB) cutoff values to classify participants as PIB+ or PIB-. They used the 
+#' See the reference below and the tacmagic walkthrough vignette. Aizenstein et
+#' al. (2008) proposed a standardized method of calculating Pittsburg Compound
+#' B (PIB) cutoff values to classify participants as PIB+ or PIB-. They used the
 #' distribution volume ratio (DVR) from several ROIs associated with amyloid 
-#' deposition. The steps are summarized below. cutoff_aiz() implements 1-3, 
-#' returning cutoff valuesfor each ROI. It can be used to dichotomize 
+#' deposition. The steps are summarized below. cutoff_aiz() implements 1-3,
+#' returning cutoff valuesfor each ROI. It can be used to dichotomize
 #' participants, with pos_anyroi().
 #' 
 #' 1. Remove outliers from a group of cognitively normal individuals. An outlier
-#' is defined as having any ROI with DVR > upper inner fence of that ROI (= 3rd 
+#' is defined as having any ROI with DVR > upper inner fence of that ROI (= 3rd
 #' quartile + (1.5 * IQR).
 #' 2. Iterate step 1 as needed until there are no more outlying participants.
 #' 3. From this subset of the group with outliers removed, the cutoff value for 
@@ -30,6 +30,9 @@
 #' @param ROIs list of variables (ROIs) to use for cutoff detection
 #' @return Cutoff values for each ROI based on the above method
 #' @family Cutoff functions
+#' @references Aizenstein HJ, Nebes RD, Saxton JA, et al. 2008. Frequent amyloid 
+#' deposition without significant cognitive impairment among the elderly. 
+#' Arch Neurol 65: 1509-1517.
 #' @examples 
 #' cutoff_aiz(fake_DVR, c("ROI1_DVR", "ROI2_DVR", "ROI3_DVR", "ROI4_DVR"))
 cutoff_aiz <- function(modelstats, ROIs) {
@@ -65,18 +68,21 @@ cutoff_aiz <- function(modelstats, ROIs) {
 
 #' Dichotomize participants based on ROI cutoff values
 #' 
-#' See references()$Aizenstein. The authors proposed a standardized method of
-#' calculating PIB+ cutoff values to classify participants as PIB+ or PIB-.
-#' They used the DVR from 7 ROIs associated with amyloid deposition. This
-#' function takes the ROI-based cutoff-values, e.g. from cutoff_aiz(), and
-#' returns a table specifying which participants are positive, i.e. which have
-#' at least one ROI greater than the cutoff.
+#' Aizenstein et al. (2008) proposed a standardized method of calculating PIB+ 
+#' cutoff values to classify participants as PIB+ or PIB-. They used the DVR 
+#' from 7 ROIs associated with amyloid deposition. This function takes the 
+#' ROI-based cutoff values, e.g. from cutoff_aiz(), and returns a table 
+#' specifying which participants are positive, i.e. which have at least one ROI
+#' greater than the cutoff.
 #' 
 #' @export
 #' @param modelstats SUVR or DVR data for group of participants from batch_tm()
 #' @param cutoff cutoffs for ROIs as from cutoff_aiz()
 #' @return data.frame of participants and positive/negative status
 #' @family Cutoff functions
+#' @references Aizenstein HJ, Nebes RD, Saxton JA, et al. 2008. Frequent amyloid 
[…]
