+++ NAMESPACE
@@ -12,7 +12,6 @@ export(load_vol)
 export(plot_ref_Logan)
 export(plot_tac)
 export(pos_anyroi)
-export(references)
 export(roi_ham_full)
 export(roi_ham_pib)
 export(roi_ham_stand)

[…]
