### 2019-02-19 82b7eed
fixed example calls to help() in walkthrough
