+++ R/logan.R
@@ -22,7 +22,7 @@
 #'@param k2prime A fixed value for k2' must be specified (e.g. 0.2)
 #'@param t_star If 0, t* will be calculated using find_t_star()
 #'@param error For find_t_star()
-#'@param method Method of inntegration, "trapz" or "integrate"
+#'@param method Method of integration, "trapz" or "integrate"
 #'@return Data frame with calculate DVRs for all ROIs
 #'@family Logan plot functions
 #'@references Logan, J., Fowler, J. S., Volkow, N. D., Wang, G.-J., 
@@ -62,7 +62,7 @@ DVR_ref_Logan <- function(tac_data, target, ref, k2prime, t_star, error=0.10,
 #'@param k2prime Required -- A fixed value for k2' must be specified (e.g. 0.2)
 #'@param t_star Required -- If 0, t* will be calculated using find_t_star()
 #'@param error For find_t_star()
-#'@param method Method of inntegration, "trapz" or "integrate"
+#'@param method Method of integration, "trapz" or "integrate"
 #'@param params Used by batch_tm (not for calling individually) to pass model
 #'              parameters
 #'@return Data frame with calculate DVRs for all ROIs
