+++ man/plot_ref_Logan.Rd
@@ -20,7 +20,7 @@ plot_ref_Logan(tac_data, target, ref, k2prime, t_star = 0, error = 0.1,
 
 \item{error}{For find_t_star()}
 
-\item{method}{Method of inntegration, "trapz" or "integrate"}
+\item{method}{Method of integration, "trapz" or "integrate"}
 }
 \value{
 No return
