### 2019-02-19 #280 @eebrown
Thank you very much @jonclayden for the thorough review and great suggestions. I will address each and report back here.
