### 2019-02-20 f3d8fe3
tac objects are now informal S3 classes; plot_tac changed to plot method, and added print method.
