+++ R/plot.R
@@ -9,12 +9,14 @@
 #' Plots time activity curves from 1 or 2 participants or groups.
 #'
 #'@export
-#'@param TACtable1 (e.g. from tac_roi() or load_tac())
-#'@param TACtable2 An optional, second TAC, to plot for comparison
+#'@param x A tac object containing time-activity curves to plot, e.g. from 
+#' tac_roi() or load_tac()
+#'@param tac2 An optional, second TAC, to plot for comparison
 #'@param ROIs A vector of ROIs to plot, names matching the TAC headers
 #'@param ymax The maximum value on the y-axis
 #'@param time "seconds" or "minutes" depending on desired x-axis, converts tac
 #'@param title A title for the plot
+#'@param ... Additional arguments
 #'@return Creates a plot
 #'@family tac functions 
 #'@examples
@@ -25,40 +27,42 @@
 #' tac <- load_tac(f_raw_tac)
 #' vol <- load_vol(f_raw_vol)
 #' AD06_tac_nc <- tac_roi(tac, vol, roi_ham_full(), merge=FALSE, PVC=FALSE)
-#' plot_tac(AD06_tac_nc, ROIs=c("frontal", "cerebellum"), title="Example Plot")
-plot_tac <- function(TACtable1, TACtable2=NULL, ROIs, ymax=25, 
-                     time="minutes", title="") {
+#' plot(AD06_tac_nc, ROIs=c("frontal", "cerebellum"), title="Example Plot")
+plot.tac <- function(x, tac2=NULL, ROIs, ymax=25, 
+                     time="minutes", title="", ...) {
   
-  if (!all(ROIs %in% names(TACtable1))) stop("ROIs are not in TACtable1")
+  tac1 <- x
 
-  if (!validate_tac(TACtable1)) stop("The 1st tac object did not validate.")
-  if (!is.null(TACtable2)) {
-    if (!validate_tac(TACtable2)) stop("The 2nd tac object did not validate.")
+  if (!all(ROIs %in% names(tac1))) stop("ROIs are not in TACtable1")
+
+  if (!validate_tac(tac1)) stop("The 1st tac object did not validate.")
+  if (!is.null(tac2)) {
+    if (!validate_tac(tac2)) stop("The 2nd tac object did not validate.")
   }
   
   if (!(time %in% c("seconds", "minutes"))) stop("Time must be \"seconds\" or 
                                                  \"minutes\"")
 
   if (time == "minutes") {
-    if (attributes(TACtable1)$time_unit == "minutes") time_conversion <- 1
-    if (attributes(TACtable1)$time_unit == "seconds") time_conversion <- 60
+    if (attributes(tac1)$time_unit == "minutes") time_conversion <- 1
+    if (attributes(tac1)$time_unit == "seconds") time_conversion <- 60
   }
 
   if (time == "seconds") {
-    if (attributes(TACtable1)$time_unit == "seconds") time_conversion <- 1
-    if (attributes(TACtable1)$time_unit == "minutes") time_conversion <- 1/60
+    if (attributes(tac1)$time_unit == "seconds") time_conversion <- 1
+    if (attributes(tac1)$time_unit == "minutes") time_conversion <- 1/60
   }
 
   # Sets up the plot using the frame start from the TAC file for the x axis
   # and converting to minutes if chosen. 
    
   plot(1,type='n',
-       xlim=c(TACtable1$start[1],
-              TACtable1$start[length(TACtable1$start)]/time_conversion),
+       xlim=c(tac1$start[1],
+              tac1$start[length(tac1$start)]/time_conversion),
               ylim=c(0,ymax),
[…]
