+++ R/utilities.R
@@ -20,6 +20,6 @@ new_table <- function(tac, varname="VALUE") {
 copy_tac_attributes <- function(origin, destination) {
   attributes(destination)$time_unit <- attributes(origin)$time_unit
   attributes(destination)$activity_unit <- attributes(origin)$activity_unit
-  attributes(destination)$tm_type <- "tac"
+  attributes(destination)$class <- c("tac", "data.frame")
   return(attributes(destination))
 }

[…]
