### 2019-02-20 4d0f181
added option to specify other colour palettes in plot
