### 2019-02-20 0d98a26
cleaned up plotting function and added ellipsis to DVR plot
