### 2019-02-20 5e169fc
Logan model now as a class and logan plot now as a method
