+++ R/batches.R
@@ -35,7 +35,7 @@
 #' tacs <- batch_load(participants, tac_file_suffix="")
 #'
 #' # Keeps only the ROIs without partial-volume correction (PMOD convention)
-#' lapply(tacs, split_pvc, FALSE)
+#' tacs <- lapply(tacs, split_pvc, FALSE)
 #' 
 #' batch <- batch_tm(tacs, models=c("SUVR", "Logan"), ref="Cerebellum_r",
 #'                   SUVR_def=c(3000,3300,3600), k2prime=0.2, t_star=23)
