+++ R/logan.R
@@ -43,11 +43,18 @@
 #'                             
 DVR_ref_Logan <- function(tac_data, target, ref, k2prime, t_star, error=0.10, 
                           method="trapz") {
-    model <- ref_Logan_lm(tac_data=tac_data, target=target, ref=ref, 
-                          k2prime=k2prime, t_star=t_star, error=error, 
-                          method=method)
-    DVR <- model$coefficients[[2]]
-    return(DVR)
+    logan_model <- ref_Logan_lm(tac_data=tac_data, target=target, ref=ref, 
+                                k2prime=k2prime, t_star=t_star, error=error, 
+                                method=method)
+    tac <- as.data.frame(tac_data)[,c("start", "end", target, ref)]
+    attributes(tac) <- copy_tac_attributes(tac_data, tac)
+    out <- list(DVR=logan_model$model$coefficients[[2]], 
+                model=logan_model$model,
+                xy=logan_model$xy,
+                t_star=logan_model$t_star,
+                tac=tac)
+    class(out) <- "ref_Logan"
+    return(out)
 }
 
 
@@ -112,113 +119,11 @@ DVR_all_ref_Logan <- function(tac_data,
                                      error=error, method=method))
         if (class(attempt) == "try-error") {
           attempt <- NA
-        }  
+        } else attempt <- attempt$DVR 
       }
         
     DVRtable[ROI, "DVR"] <- attempt
   }
 
   return(DVRtable)
-}
-
-
-## Helper functions------------------------------------------------------------
-
-# The non-invasive reference Logan method
-#' @noRd
-ref_Logan_xy <- function(tac, target, ref, k2prime, method) {
-  
-  if (!(validate_tac(tac))) stop("Invalid tac object provided.")
-
-  mid_time <- (tac$start + tac$end) / 2
-  if (attributes(tac)$time_unit == "seconds") {
-    mid_time <- mid_time / 60 # needed because k2' is units 1/min (not seconds)
-  }
-  
-  # Derive functions for TACs by interpolation.
-  target_tac <- approxfun(x=mid_time, y=tac[,target], method="linear", rule=2)
-  ref_tac <- approxfun(x=mid_time, y=tac[,ref], method = "linear", rule=2)
-
-  if (!is.null(k2prime)) k2r <- (tac[,ref] / k2prime) else k2r <- 0
-
-  if (method == "trapz") {
-    frames <- seq_along(mid_time)
-    yA <- vapply(frames, FUN=wrap_auc, FUN.VALUE=0, x=mid_time, y=tac[,target])
-    xA <- vapply(frames, FUN=wrap_auc, FUN.VALUE=0, x=mid_time, y=tac[,ref]) + 
-                         k2r
-  } else if (method == "integrate") {
-    yA <- vapply(mid_time, FUN=wrap_integrate, FUN.VALUE=0, lower=mid_time[1], 
-                 f=target_tac)
-    xA <- vapply(mid_time, FUN=wrap_integrate, FUN.VALUE=0, lower=mid_time[1], 
-                 f=ref_tac) + k2r
-  }
-
-  y <- yA / tac[,target]
-  x <- xA / tac[,target]
-
-  output <- data.frame(x,y)
-
-  return(output)
-}
-
-
-# The non-invasive reference Logan method -- linear model starting from t*
-#' @noRd
-ref_Logan_lm <- function(tac_data, target, ref, k2prime, t_star, error, 
-                         method) {
-    
-  xy <- ref_Logan_xy(tac_data, target=target, ref=ref, k2prime=k2prime,
-                     method=method)
-  
-  x <- xy$x
-  y <- xy$y
-  
-  if (t_star == 0) {
-    t_star <- find_t_star(x, y, error=error)
[…]
