+++ man/batch_tm.Rd
@@ -45,7 +45,7 @@ participants <- c(system.file("extdata", "AD06.tac", package="tacmagic"),
 tacs <- batch_load(participants, tac_file_suffix="")
 
 # Keeps only the ROIs without partial-volume correction (PMOD convention)
-lapply(tacs, split_pvc, FALSE)
+tacs <- lapply(tacs, split_pvc, FALSE)
 
 batch <- batch_tm(tacs, models=c("SUVR", "Logan"), ref="Cerebellum_r",
                   SUVR_def=c(3000,3300,3600), k2prime=0.2, t_star=23)

[…]
