### 2019-02-20 6728c31
updated batch_tm() to use the ellipsis argument for cleaner and more flexible code.
