### 2019-02-20 0c1ffb1
validate_ref_Logan_params() added
