### 2019-02-21 fab5007
additional tests, and fix for split_pvc()
