### 2019-02-21 8471eef
update to paper.md for typos and formatting
