### 2019-02-21 #280 @eebrown
Thanks for clarifying @jonclaydon. I can do this and it makes sense. In the future I may have to switch to the ellipsis rather than the specific parameters as other DVR models need different parameters.
