### 2019-02-21 #280 @jonclayden
Sure, that makes sense – you could keep `DVR_ref_Logan()` as a method-specific function and create `dvr()` as a wrapper with an ellipsis argument, for example.

Everything else looks good, although you may want to implement an indexing method, `[.tac`, or make your `print` method slightly more defensive so that you don't need to coerce back to a data frame when subsetting.
