### 2019-02-22 f5cb2f9
changed print.tac to summary.tac, created as.tac() function and testing for it; added to walkthrough vignette
