+++ doc/walkthrough.R
@@ -14,9 +14,9 @@ filename <- system.file("extdata", "AD06.tac", package="tacmagic")
 AD06_tac <- load_tac(filename, format="PMOD")
 
 ## ------------------------------------------------------------------------
-print(AD06_tac) 
+summary(AD06_tac) 
 
-as.data.frame(AD06_tac)[1:5,1:5] # the first 5 frames of the first 3 ROIs
+AD06_tac[1:5,1:5] # the first 5 frames of the first 3 ROIs
 
 ## ------------------------------------------------------------------------
 filename_acq <- system.file("extdata", "AD06.acqtimes", package="tacmagic")
@@ -32,7 +32,13 @@ f_magia <- system.file("extdata", "AD06_tac_magia.mat", package="tacmagic")
 AD06_tac_magia <- load_tac(f_magia, format="magia", 
                            time_unit="seconds", activity_unit="kBq/cc")
 
-as.data.frame(AD06_tac_magia)[1:5,1:5]
+AD06_tac_magia[1:5,1:5]
+
+## ------------------------------------------------------------------------
+manual <- data.frame(start=c(0:4), end=c(2:6), ROI1=c(10.1:14.2), ROI2=c(11:15))
+manual_tac <- as.tac(manual, time_unit="minutes", activity_unit="kBq/cc")
+
+summary(manual_tac)
 
 ## ------------------------------------------------------------------------
 AD06_volume <- load_vol(filename_voistat, format="voistat")
@@ -46,7 +52,7 @@ AD06 <- tac_roi(tac=AD06_tac,           # The tac file we loaded above.
                 PVC=F                   # to use _C ROIs (PMOD convention)            
                 )
 
-as.data.frame(AD06)[1:5,1:5]
+AD06[1:5,1:5]
 
 ## ---- fig.show='hold', fig.height=4.5, fig.width=6.5, fig.align='center'----
 plot(AD06,                                                    # tac data
