+++ man/load_voistat.Rd
@@ -29,7 +29,7 @@ f <- system.file("extdata", "AD06_BPnd_BPnd_Logan.voistat",
 vs <- load_voistat(f, ROI_def=roi_ham_pib(), model="Logan")
 }
 \seealso{
-Other Loading functions: \code{\link{load_tac}},
-  \code{\link{load_vol}}
+Other Loading functions: \code{\link{as.tac}},
+  \code{\link{load_tac}}, \code{\link{load_vol}}
 }
 \concept{Loading functions}

[…]
