### 2019-02-22 090c0af
minor fixes
