### 2019-02-22 578a89b
added wrapper function dvr()
