+++ man/plot.ref_Logan.Rd
@@ -28,6 +28,6 @@ This plots the non-invasive Logan plot.
 }
 \seealso{
 Other Logan plot functions: \code{\link{DVR_all_ref_Logan}},
-  \code{\link{DVR_ref_Logan}}
+  \code{\link{DVR_ref_Logan}}, \code{\link{dvr}}
 }
 \concept{Logan plot functions}
