+++ NAMESPACE
@@ -11,6 +11,7 @@ export(batch_load)
 export(batch_tm)
 export(batch_voistat)
 export(cutoff_aiz)
+export(dvr)
 export(load_tac)
 export(load_voistat)
 export(load_vol)

[…]
