+++ paper.bib
@@ -9,7 +9,7 @@
 
 @misc{tpcclib,
 author = {Oikonen, V.},
-year = {n.d.},
+year = {2018},
 title = {tpcclib: Command-line tools to processing and analyzing data collected in Turku PET Centre},
 howpublished = {\url{https://gitlab.utu.fi/vesoik/tpcclib}},
 publisher = {Turku PET Centre, Turku, Finland}
