### 2019-02-25 27dacbf
added rOpenSci footer to readme
