### 2019-02-25 #280 @melvidoni
@eebrown I don't know why, but I don't see the box ticked on the opening post for the issue. Nevermind this, I apologize for the misunderstanding.

I have added [three additional to-dos on the approval comment](https://github.com/ropensci/software-review/issues/280#issuecomment-466962697) and I need you to tackle those as well. Do everything before submitting to JOSS, please. You should have already received an invitation for the team.
