### 2019-02-25 16aad9b
added a rOpenSci reviewer to DESCRIPTION
