### 2019-02-25 3c46aa2
added 2nd rOpenSci reviewer to DESCRIPTION
