### 2019-02-25 0d5216e
Incremented version number to 0.2.0 post-review; regenerated documentation, vignette, site, and meta
