### 2019-02-25 d770b9b
update installation instructions to new repo location
