### 2019-02-26 #280 @eebrown
Thanks very much @melvidoni; it works now, and a DOI has been assigned.

[![DOI](https://zenodo.org/badge/131427691.svg)](https://zenodo.org/badge/latestdoi/131427691)

On the other hand I don't seem to be able to edit your comment, so I've copied the post-approval checklist below. Essentially I am at the last step, about to submit to JOSS. As I would also like to submit to CRAN, does the order of submitting to JOSS/CRAN matter from your perspective?

> To-dos:
>
> * [x]  Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do.
> * [x]  Add the rOpenSci footer to the bottom of your README
>
> ```
> [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
> ```
> * [x]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`.
> * [x]  We're starting to roll out software metadata files to all ropensci packages via the Codemeta initiative, see https://github.com/ropensci/codemetar/#codemetar for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package.
> * [x]  Activate Zenodo watching the repo
> * [x]  Tag and create a release so as to create a Zenodo version and DOI
> * [ ]  Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

 * [x] Reviewers acknowledged in DESCRIPTION
