### 2019-02-26 #280 @melvidoni
Thanks, @eebrown, now:

1) Submit package via http://joss.theoj.org/papers/new

2) Watch for the paper to pop up at http://joss.theoj.org/papers, then add the following comment to the submission thread:

```
This submission has been accepted to rOpenSci. The review thread can be found at [LINK TO SOFTWARE REVIEW ISSUE]
```

If everything is OK, I am closing the issue.
