### 2019-02-26 494f8f7
minor spelling/typo corrections; no code changes; docs rebuilt
