+++ man/dvr.Rd
@@ -33,7 +33,7 @@ all TACs in a supplied tac file. It uses DVR_ref_Logan if a target ROI is
 specified, otherwise will calculate DVR for all ROIs with DVR_ref_all_Logan()
 }
 \details{
-For other model paramters, directly call DVR_ref_Logan().
+For other model parameters, directly call DVR_ref_Logan().
 }
 \examples{
 f <- system.file("extdata", "AD06.tac", package="tacmagic")
