### 2019-02-26 09ca0f4
specified tests to skip_on_cran  due to platform-specific plotting differences; typo in DESCRIPTION
