### 2019-02-26 9e9eab4
update to DESCRIPTION based on feedback from CRAN submission; typos
