+++ vignettes/walkthrough.Rmd
@@ -23,13 +23,13 @@ knitr::opts_chunk$set(
 
 [Positron emission tomography](https://en.wikipedia.org/wiki/Positron_emission_tomography) (PET) is a research and clinical imaging modality that uses radioactive tracers that bind to target molecules of interest. A PET scanner identifies the tracer location by virtue of the tracer's radioactive decay, providing information to determine the location of the target in the body. As the spatial resolution of PET is relatively poor, analysis is frequently combined with higher resolution imaging such as magnetic resonance imaging (MRI), which can be spatially co-registered to the PET image. Subsequently, radiotracer activity (over time) can be identified by spatial region of interest (ROI).
 
-An image anaylsis pipeline is required to extract regional time activity curves (tacs) from a dynamic PET image. There are various pipelines available including widely-used commercial solutions (e.g. [PMOD](https://www.pmod.com/web/)) and newer open-source options (e.g. [magia](http://aivo.utu.fi/magia/)). Pipelines generally implement the following steps:
+An image analysis pipeline is required to extract regional time activity curves (TACs) from a dynamic PET image. There are various pipelines available including widely-used commercial solutions (e.g. [PMOD](https://www.pmod.com/web/)) and newer open-source options (e.g. [magia](http://aivo.utu.fi/magia/)). Pipelines generally implement the following steps:
 
 * Dynamic PET pre-processing (e.g. motion correction, decay-correction)
 * PET image co-registration with structural MRI
 * MRI segmentation and normalization to atlas
 
-Various pipelines save tac, volume and related data in various formats. This package enables the loading and analysis of tac and ROI volume data from image analysis pipelines for further analysis in R. 
+Various pipelines save TAC, volume and related data in various formats. This package enables the loading and analysis of TAC and ROI volume data from image analysis pipelines for further analysis in R. 
 
 ## Vignette data
 
@@ -41,24 +41,24 @@ There are two approaches to using the **tacmagic** package to analyze PET time-a
 
 ### Data loading
 
-Time-activity curve (TAC) data is loaded via `load_tac()`, which is a wrapper for format-specific functions. To specify which file format the tac data is stored as, use the `format` parameter. Supported formats can be viewed in `help(load_tac)`.
+Time-activity curve (TAC) data is loaded via `load_tac()`, which is a wrapper for format-specific functions. To specify which file format the TAC data is stored as, use the `format` parameter. Supported formats can be viewed in `help(load_tac)`.
 
 The minimal amount of information required is the TAC data for one or more ROI, including the start and stop times of each frame, the time units and the activity units. This information may be in 1 or more files depending on the format and software that created it. 
 
 For example, PMOD's `.tac` files contain all of the information, but the TAC .voistat files do not contain start and stop times, but this information could be specified using a `.acqtimes` file. Support is also available for DFT format, which contains both TAC and volume data.
 
-We processed the PIB PET and T1 MRI data with the [PMOD](https://www.pmod.com/) PNEURO software suite to produce a `.tac` file with tacs for all ROIs in the Hammer's atlas. The `.tac` file can be loaded with `load_tac()`:
+We processed the PIB PET and T1 MRI data with the [PMOD](https://www.pmod.com/) PNEURO software suite to produce a `.tac` file with TACs for all ROIs in the Hammer's atlas. The `.tac` file can be loaded with `load_tac()`:
 
 ```{r}
 # Filename is a character string of the file's path on your computer.
 filename <- system.file("extdata", "AD06.tac", package="tacmagic")
[…]
