+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: tacmagic
 Type: Package
-Title: PET Analysis in R
+Title: Positron Emission Tomography Time-Activity Curve Analysis
 Version: 0.2.0
 Authors@R: c(person("Eric", "Brown", 
                     role = c("aut", "cre"),
@@ -16,18 +16,20 @@ Authors@R: c(person("Eric", "Brown",
                     role = c("rev"),
                     comment = c(ORCID = "0000-0003-2576-4544", 
                     "Brandon Hurr reviewed the package for ropensci, see <https://github.com/ropensci/software-review/issues/280>")))
-Description: To facilitate analysis of positron emission tomography (PET) time
-    activity curve (TAC) data, and to encourage open science and replicability,
-    this package supports data loading and analysis of multiple TAC file 
-    formats. Functions are available to analyze loaded TAC data for individual 
-    participants or in batches. Major functionality includes weighted TAC 
-    merging by region of interest (ROI), calculating models including SUVR and
-    DVR, basic plotting functions and calculation of cut-off values. Please see
-    the walkthrough vignette for a detailed overview of tacmagic functions.
+Description: To facilitate the analysis of positron emission tomography (PET) 
+    time activity curve (TAC) data, and to encourage open science and 
+    replicability, this package supports data loading and analysis of multiple 
+    TAC file formats. Functions are available to analyze loaded TAC data for 
+    individual participants or in batches. Major functionality includes weighted
+    TAC merging by region of interest (ROI), calculating models including SUVR
+    and DVR (Logan et al. 1996 <doi:10.1097/00004647-199609000-00008>), basic 
+    plotting functions and calculation of cut-off values (Aizenstein et al. 2008
+    <doi:10.1001/archneur.65.11.1509>). Please see the walkthrough vignette for
+    a detailed overview of tacmagic functions.
 Depends: R (>= 3.4)
 License: GPL-3
-URL: https://github.com/eebrown/PET
-BugReports: https://github.com/eebrown/PET/issues
+URL: https://github.com/ropensci/tacmagic
+BugReports: https://github.com/ropensci/tacmagic
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
